+++ modules/router/src/test/collateral.spec.ts
@@ -11,7 +11,7 @@ import {
 import Sinon from "sinon";
 import { BigNumber } from "@ethersproject/bignumber";
 import { AddressZero } from "@ethersproject/constants";
-import { ChainError, INodeService, Result, UpdateType } from "@connext/vector-types";
+import { ChainError, INodeService, IVectorChainReader, Result, UpdateType } from "@connext/vector-types";
 import { parseEther } from "@ethersproject/units";
 
 import { getConfig } from "../config";
@@ -72,7 +72,7 @@ describe(testName, () => {
         AddressZero,
         routerPublicIdentifier,
         node as INodeService,
-        chainReader,
+        chainReader as IVectorChainReader,
         log,
         transferAmount.toString(),
       );
@@ -87,7 +87,7 @@ describe(testName, () => {
         AddressZero,
         routerPublicIdentifier,
         node as INodeService,
-        chainReader,
+        chainReader as IVectorChainReader,
         log,
         transferAmount.toString(),
       );
@@ -108,7 +108,7 @@ describe(testName, () => {
         AddressZero,
         routerPublicIdentifier,
         node as INodeService,
-        chainReader,
+        chainReader as IVectorChainReader,
         log,
         transferAmount.toString(),
       );
@@ -129,7 +129,7 @@ describe(testName, () => {
         AddressZero,
         routerPublicIdentifier,
         node as INodeService,
-        chainReader,
+        chainReader as IVectorChainReader,
         log,
         transferAmount.toString(),
       );
@@ -141,8 +141,11 @@ describe(testName, () => {
       const { channel } = createTestChannelState(UpdateType.deposit, {
         alice: mkAddress("0xaaa"),
         aliceIdentifier: routerPublicIdentifier,
+        bob: mkAddress("0xbbb"),
         assetIds: [AddressZero],
         balances: [{ to: [mkAddress("0xaaa"), mkAddress("0xbbb")], amount: ["0", "0"] }],
+        processedDepositsA: ["0"],
+        processedDepositsB: ["0"],
       });
       node.getStateChannel.onCall(0).resolves(Result.ok(channel));
       node.getStateChannel.onCall(1).resolves(
@@ -161,7 +164,7 @@ describe(testName, () => {
         AddressZero,
         routerPublicIdentifier,
         node as INodeService,
-        chainReader,
+        chainReader as IVectorChainReader,
         log,
         transferAmount.toString(),
       );
@@ -183,6 +186,8 @@ describe(testName, () => {
         aliceIdentifier: routerPublicIdentifier,
         assetIds: [AddressZero],
         balances: [{ to: [mkAddress("0xaaa"), mkAddress("0xbbb")], amount: ["0", "0"] }],
+        processedDepositsA: ["0"],
+        processedDepositsB: ["0"],
       });
       const profile = { ...ethProfile, target: "0" };
       getRebalanceProfile.returns(Result.ok(profile));
@@ -205,7 +210,7 @@ describe(testName, () => {
         AddressZero,
         routerPublicIdentifier,
         node as INodeService,
-        chainReader,
+        chainReader as IVectorChainReader,
         log,
         transferAmount.toString(),
       );
@@ -236,7 +241,7 @@ describe(testName, () => {
         AddressZero,
         routerPublicIdentifier,
         node as INodeService,
-        chainReader,
+        chainReader as IVectorChainReader,
         log,
       );
       expect(res.getError()).to.be.undefined;
@@ -251,6 +256,8 @@ describe(testName, () => {
         aliceIdentifier: routerPublicIdentifier,
         assetIds: [AddressZero],
         balances: [{ to: [mkAddress("0xaaa"), mkAddress("0xbbb")], amount: ["0", "0"] }],
+        processedDepositsA: ["0"],
+        processedDepositsB: ["0"],
       });
       node.getStateChannel.onFirstCall().resolves(Result.ok(channel));
       node.getStateChannel.resolves(
@@ -269,7 +276,7 @@ describe(testName, () => {
         AddressZero,
         routerPublicIdentifier,
         node as INodeService,
-        chainReader,
+        chainReader as IVectorChainReader,
         log,
       );
       expect(res.getError()).to.be.undefined;
@@ -311,7 +318,7 @@ describe(testName, () => {
         AddressZero,
         routerPublicIdentifier,
         node as INodeService,
-        chainReader,
+        chainReader as IVectorChainReader,
         log,
       );
       expect(res.getError()).to.be.undefined;
@@ -348,7 +355,7 @@ describe(testName, () => {
         AddressZero,
         routerPublicIdentifier,
         node as INodeService,
-        chainReader,
+        chainReader as IVectorChainReader,
         log,
       );
       expect(res.getError()!.message).to.be.eq(CollateralError.reasons.UnableToReclaim);
@@ -374,6 +381,8 @@ describe(testName, () => {
             amount: [routerBalance.toString(), "0"],
           },
         ],
+        processedDepositsA: [routerBalance.toString()],
+        processedDepositsB: ["0"],
       });
       const profile = { ...ethProfile, target: "0" };
       getRebalanceProfile.returns(Result.ok(profile));
@@ -390,7 +399,7 @@ describe(testName, () => {
         AddressZero,
         routerPublicIdentifier,
         node as INodeService,
-        chainReader,
+        chainReader as IVectorChainReader,
         log,
       );
       expect(res.getError()).to.be.undefined;
@@ -421,7 +430,7 @@ describe(testName, () => {
         AddressZero,
         routerPublicIdentifier,
         node as INodeService,
-        chainReader,
+        chainReader as IVectorChainReader,
         log,
       );
       expect(res.getError()).to.be.undefined;
@@ -443,7 +452,7 @@ describe(testName, () => {
         AddressZero,
         routerPublicIdentifier,
         node as INodeService,
-        chainReader,
+        chainReader as IVectorChainReader,
         log,
       );
       expect(res.getError()!.message).to.be.eq(CollateralError.reasons.UnableToGetRebalanceProfile);
@@ -460,7 +469,7 @@ describe(testName, () => {
         AddressZero,
         routerPublicIdentifier,
         node as INodeService,
-        chainReader,
+        chainReader as IVectorChainReader,
         log,
       );
       expect(res.getError()!.message).to.be.eq(CollateralError.reasons.ProviderNotFound);
@@ -477,7 +486,7 @@ describe(testName, () => {
         AddressZero,
         routerPublicIdentifier,
         node as INodeService,
-        chainReader,
+        chainReader as IVectorChainReader,
         log,
       );
       expect(res.getError()!.message).to.be.eq(CollateralError.reasons.ProviderNotFound);
@@ -494,7 +503,7 @@ describe(testName, () => {
         AddressZero,
         routerPublicIdentifier,
         node as INodeService,
-        chainReader,
+        chainReader as IVectorChainReader,
         log,
       );
       expect(res.getError()!.message).to.be.eq(CollateralError.reasons.CouldNotGetOnchainDeposits);
@@ -510,6 +519,8 @@ describe(testName, () => {
               amount: ["0", "0"],
             },
           ],
+          processedDepositsA: ["0"],
+          processedDepositsB: ["0"],
         });
         const requestedAmount = BigNumber.from(ethProfile.target).add(10000);
         node.getStateChannel.resolves(
@@ -528,7 +539,7 @@ describe(testName, () => {
           AddressZero,
           routerPublicIdentifier,
           node as INodeService,
-          chainReader,
+          chainReader as IVectorChainReader,
           log,
           requestedAmount.toString(),
         );
@@ -553,6 +564,8 @@ describe(testName, () => {
               amount: ["0", "0"],
             },
           ],
+          processedDepositsA: ["0"],
+          processedDepositsB: ["0"],
         });
         const requestedAmount = BigNumber.from(ethProfile.target).sub(10000);
         node.getStateChannel.resolves(
@@ -571,7 +584,7 @@ describe(testName, () => {
           AddressZero,
           routerPublicIdentifier,
           node as INodeService,
-          chainReader,
+          chainReader as IVectorChainReader,
           log,
           requestedAmount.toString(),
         );
@@ -596,6 +609,8 @@ describe(testName, () => {
               amount: ["0", "0"],
             },
           ],
+          processedDepositsA: ["0"],
+          processedDepositsB: ["0"],
         });
         node.getStateChannel.resolves(
           Result.ok({
@@ -613,7 +628,7 @@ describe(testName, () => {
           AddressZero,
           routerPublicIdentifier,
           node as INodeService,
-          chainReader,
+          chainReader as IVectorChainReader,
           log,
         );
         expect(res.isError).to.be.false;
@@ -640,7 +655,7 @@ describe(testName, () => {
           AddressZero,
           routerPublicIdentifier,
           node as INodeService,
-          chainReader,
+          chainReader as IVectorChainReader,
           log,
         );
         expect(res.isError).to.be.false;
@@ -671,7 +686,7 @@ describe(testName, () => {
           AddressZero,
           routerPublicIdentifier,
           node as INodeService,
-          chainReader,
+          chainReader as IVectorChainReader,
           log,
         );
         expect(res.isError).to.be.false;
@@ -688,6 +703,8 @@ describe(testName, () => {
               amount: ["0", "0"],
             },
           ],
+          processedDepositsA: ["0"],
+          processedDepositsB: ["0"],
         });
         const requestedAmount = BigNumber.from(ethProfile.target).add(10000);
         const profile = { ...ethProfile, target: "0" };
@@ -708,7 +725,7 @@ describe(testName, () => {
           AddressZero,
           routerPublicIdentifier,
           node as INodeService,
-          chainReader,
+          chainReader as IVectorChainReader,
           log,
           requestedAmount.toString(),
         );
@@ -750,7 +767,7 @@ describe(testName, () => {
           AddressZero,
           routerPublicIdentifier,
           node as INodeService,
-          chainReader,
+          chainReader as IVectorChainReader,
           log,
           requestedAmount.toString(),
         );
@@ -761,7 +778,7 @@ describe(testName, () => {
           AddressZero,
           routerPublicIdentifier,
           node as INodeService,
-          chainReader,
+          chainReader as IVectorChainReader,
           log,
           requestedAmount.toString(),
         );
@@ -770,6 +787,56 @@ describe(testName, () => {
         expect(node.sendDepositTx.callCount).to.be.eq(2);
         expect(node.reconcileDeposit.callCount).to.be.eq(2);
       });
+    });
+
+    it("should collateralize difference between owed and target", async () => {
+      const { channel } = createTestChannelState(UpdateType.deposit, {
+        aliceIdentifier: routerPublicIdentifier,
+        alice: mkAddress("0xaaaaa"),
+        bob: mkAddress("0xbbbbcccc"),
+        processedDepositsA: ["0"],
+        assetIds: [AddressZero],
+        balances: [
+          {
+            amount: ["0", "0"],
+            to: [mkAddress("0xaaaaa"), mkAddress("0xbbbbcccc")],
+          },
+        ],
+      });
+      const requestedAmount = parseEther("0.001");
+      chainReader.getTotalDepositedA.resolves(Result.ok(requestedAmount.sub(10)));
+      node.getStateChannel.resolves(
+        Result.ok({
+          ...channel,
+          balances: [
+            {
+              to: [mkAddress("0xaaa"), mkAddress("0xbbb")],
+              amount: [requestedAmount.toString(), "0"],
+            },
+          ],
+        }),
+      );
+
+      const res = await requestCollateral(
+        channel,
+        AddressZero,
+        routerPublicIdentifier,
+        node as INodeService,
+        chainReader as IVectorChainReader,
+        log,
+        requestedAmount.toString(),
+      );
+
+      expect(node.sendDepositTx.getCall(0).args[0]).to.deep.eq({
+        amount: "10",
+        assetId: AddressZero,
+        chainId: channel.networkContext.chainId,
+        channelAddress: channel.channelAddress,
+        publicIdentifier: routerPublicIdentifier,
+      });
+      expect(res.isError).to.be.false;
+      expect(node.sendDepositTx.callCount).to.be.eq(1);
+      expect(node.reconcileDeposit.callCount).to.be.eq(1);
     });
   });
 });